consolidated the PermissionServiceTest since it'll always be authenticated
